LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/popup-same-origin-with-same-origin.https-expected.txt


FAIL same-origin document opening popup to https://localhost:9443 with COOP: "" assert_equals: name expected "" but got "same-origin_to_SAME_ORIGIN_"
FAIL same-origin document opening popup to https://localhost:9443 with COOP: "unsafe-none" assert_equals: name expected "" but got "same-origin_to_SAME_ORIGIN_unsafe-none"
FAIL same-origin document opening popup to https://localhost:9443 with COOP: "jibberish" assert_equals: name expected "" but got "same-origin_to_SAME_ORIGIN_jibberish"
FAIL same-origin document opening popup to https://localhost:9443 with COOP: "same-site" assert_equals: name expected "" but got "same-origin_to_SAME_ORIGIN_same-site"
FAIL same-origin document opening popup to https://localhost:9443 with COOP: "same-site unsafe-allow-outgoing" assert_equals: name expected "" but got "same-origin_to_SAME_ORIGIN_same-site-unsafe-allow-outgoing"
FAIL same-origin document opening popup to https://localhost:9443 with COOP: "same-origin unsafe-allow-outgoing" assert_equals: name expected "" but got "same-origin_to_SAME_ORIGIN_same-origin-unsafe-allow-outgoing"
PASS same-origin document opening popup to https://localhost:9443 with COOP: "same-origin"
FAIL same-origin document opening popup to https://localhost:9443 with COOP: "same-origin-allow-popups" assert_equals: name expected "" but got "same-origin_to_SAME_ORIGIN_same-origin-allow-popups"
